Extracted getting the exe's location to a static variable in PathsBuilder. Reasoning below:
I found out with another application that Assembly.GetExecutingAssembly().Location would return null if the project is published as a single file.
To avoid similar issues, I've centralized these uses to a single method(to be able to edit it easily) and used a .net6.0-specific method that will work even in that case.